requirements.txt
asgiref==3.5.2
Django==3.2
django-taggit==3.0.0
djangorestframework==3.13.1
intervaltree==3.1.0
networkx==2.8.2
Pillow==9.1.1
pytz==2022.1
sortedcontainers==2.4.0
sqlparse==0.4.2
